+++ src/main/java/pro/sky/java/course3/task2/repositories/StudentRepository.java
@@ -7,6 +7,6 @@ import java.util.Collection;
 import java.util.List;
 
 @Repository
-public interface StudentRepository extends JpaRepository<Student, Long>{
-    List<Student> findByAge(int age);
+public interface StudentRepository extends JpaRepository<Student, Integer>{
+    List<Student> findByAge(Integer age);
 }
